// src/bot/common/enums/conversation-steps.ts
export enum ConversationSteps {
    START = 'START',
    LANGUAGE_SELECTION = 'LANGUAGE_SELECTION',
    ASKING_PHONE = 'ASKING_PHONE',
    VALIDATION_PHONE = 'VALIDATION_PHONE',
    PHONE_VERIFICATION = 'PHONE_VERIFICATION',
    MAIN_MENU = 'MAIN_MENU',
    PURCHASES = 'PURCHASES',
    PAYMENTS = 'PAYMENTS',
    COMPLETED = 'COMPLETED'
}